반 정도는 된 것 같기도...?

+++ DX12App/stdafx.h
@@ -522,7 +522,7 @@ namespace Vector3
 {
 	inline float Distance(const XMFLOAT3& v1, const XMFLOAT3& v2)
 	{
-		return sqrt(pow(v1.x - v2.x, 2) + pow(v1.y - v2.y, 2) + pow(v1.z - v2.z, 2));
+		return (float)sqrt(pow(v1.x - v2.x, 2) + pow(v1.y - v2.y, 2) + pow(v1.z - v2.z, 2));
 	}
 
 	inline XMFLOAT3 btVectorToXM(const btVector3& v)
